% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCboot.R
\name{BCboot}
\alias{BCboot}
\title{Apply Bias Correction to Bootstrap Estimates}
\usage{
BCboot(thetastar, thetahat, bounds = c(0.1, 0.9))
}
\arguments{
\item{thetastar}{a vector of bootstrap estimates.}

\item{thetahat}{a point estimate from original data.}

\item{bounds}{a vector of lower and upper limits to handle extremely biased
cases.}
}
\value{
Vector of bias-corrected bootstrap estimates.
}
\description{
Apply bias correction to bootstrap estimates.
}
\note{
BCa with zero acceleration, based on \code{bcanon} in package \pkg{bootstrap}
by Tibshirani.

See Efron and Tibshirani (1993, pp. 184-186), Gavaris and Van Eeckhaute
(1998, p.10), and Gavaris (1999, p. 47).
}
