% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{bench}
\alias{bench}
\title{Difference between Two Times}
\usage{
bench(before, after)
}
\arguments{
\item{before}{a string like \code{"7:58:32"}.}

\item{after}{a string like \code{"8:55:09"}.}
}
\value{
Seconds as a number.
}
\description{
Calculate difference between two times in seconds.
}
