% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevholt.h.R
\name{bevholt.h}
\alias{bevholt.h}
\title{Beverton-Holt Recruitment}
\usage{
bevholt.h(R0, h, M, mat, w, B)
}
\arguments{
\item{R0}{average virgin recruitment.}

\item{h}{steepness.}

\item{M}{natural mortality rate.}

\item{mat}{maturity at age.}

\item{w}{weight at age.}

\item{B}{spawning biomass.}
}
\value{
Recruitment as a number or vector.
}
\description{
Calculate Beverton-Holt recruitment (Francis formulation).
}
\note{
Formulation from Francis (1992, p. 929), where steepness is relative
recruitment at \code{0.2*B0}.
}
\seealso{
\code{\link{bevholt}}, \code{\link{bevholt.Rmax}}.
}
