% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevholt.R
\name{bevholt}
\alias{bevholt}
\title{Beverton-Holt Recruitment}
\usage{
bevholt(S, a, b)
}
\arguments{
\item{S}{spawning biomass.}

\item{a}{initial slope.}

\item{b}{carrying capacity.}
}
\value{
Recruitment as a number or vector.
}
\description{
Calculate Beverton-Holt recruitment.
}
\note{
Formulation from Hilborn and Walters (1992, Eq. 7.5.3, p. 258).

Scale the y axis by multiplying \code{a} and \code{b} by same amount.
}
