% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevholt.Rmax.R
\name{bevholt.Rmax}
\alias{bevholt.Rmax}
\title{Beverton-Holt Recruitment}
\usage{
bevholt.Rmax(Rmax, S50, S)
}
\arguments{
\item{Rmax}{maximum recruitment.}

\item{S50}{stock size that gives 1/2 \code{Rmax}.}

\item{S}{spawning biomass.}
}
\value{
Recruitment as a number or vector.
}
\description{
Calculate Beverton-Holt recruitment (Haddon formulation).
}
\note{
Formulation from Haddon (2001, Eq. 9.2, p. 252).

Scale the y axis with \code{Rmax}.
}
