% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{Univariate Optimization}
\usage{
bisection(f, interval, tol = 1e-08, hessian = FALSE, ...)
}
\arguments{
\item{f}{a function to minimize, by varying the first argument.}

\item{interval}{a vector containing the lower and upper limits of the
interval to search over.}

\item{tol}{the max difference between the optimal \code{x} and returned
\code{x}.}

\item{hessian}{whether to return the Hessian matrix, standard error, and 95\%
confidence interval.}

\item{\dots}{passed to the \code{f} function.}
}
\value{
List containing \code{x} (optimal value), \code{f} (function value), and
\code{n} (number of function evaluations).
}
\description{
Univariate optimization.
}
\note{
The number of function evaluations (\code{n}) is two times the number of
loops performed.
}
