% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{book}
\alias{book}
\title{Bind PDF Book}
\usage{
book(directory, merge = TRUE, redistill = TRUE)
}
\arguments{
\item{directory}{where the PDF files are.}

\item{merge}{whether to merge the chapters - otherwise return \code{cmd}.}

\item{redistill}{whether to redistill \file{out.pdf} (pdftk output) to
\file{_out.pdf} (Ghostscript output).}
}
\value{
Shell command string used to merge PDF files with \code{pdftk.}
}
\description{
Bind book from many PDF files.
}
\note{
The PDF files should be named something like \file{0.pdf 1.pdf 27.pdf 42.pdf
... 397.pdf 413a.pdf 413b.pdf}.

Before this function is called, unnecessary pages can be peeled out of
\file{0.pdf}, and the last file often needs splitting, e.g. \file{413.pdf}
into \file{413a.pdf} and \file{413b.pdf}.

Three files are written to directory: \file{_} (blank page of right size),
\file{out.pdf} (book merged with pdftk), and \file{_out.pdf} (\file{out.pdf}
redistilled with Ghostscript).

After this function is called, one should check \file{out.pdf} to see whether
pages are correctly numbered, and \file{out_.pdf} to see whether the
Ghostscript-redistilled document is smaller while retaining intact images and
no red borders around links
}
\section{Warning}{

The final Ghostscript redistillation is launched with \code{wait = FALSE}, so
the best way to see if it is finished is to check if any Ghostscript
processes are active.
}

