% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\alias{bubbleplot.default}
\alias{bubbleplot.formula}
\title{Bubble Plot}
\usage{
bubbleplot(x, ...)

\method{bubbleplot}{default}(
  x,
  y,
  z,
  std = TRUE,
  pow = 0.5,
  add = FALSE,
  rev = FALSE,
  type = "p",
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  pch = c(16, 1),
  cex.points = 1,
  col = "black",
  bg = par("bg"),
  ...
)

\method{bubbleplot}{formula}(formula, data, subset, na.action = NULL, ...)
}
\arguments{
\item{x}{a vector of values for the horizontal axis, or a data frame
containing \code{x}, \code{y}, and \code{z} in that order.}

\item{\dots}{passed to \code{bubbleplot.*} methods.}

\item{y}{a vector of values for the vertical axis.}

\item{z}{a vector of values determining the bubble sizes.}

\item{std}{whether to standardize \code{z} by dividing with \code{mean(z)}.}

\item{pow}{a power coefficient for the bubble sizes (\code{0.5} is
\code{sqrt}).}

\item{add}{whether to add bubbles to existing plot.}

\item{rev}{whether to reverse the y axis.}

\item{type}{passed to \code{points}.}

\item{ylim}{passed to \code{plot}.}

\item{xlab}{passed to \code{plot}.}

\item{ylab}{passed to \code{plot}.}

\item{pch}{passed to \code{points}.}

\item{cex.points}{scales all bubble sizes.}

\item{col}{passed to \code{points}.}

\item{bg}{passed to \code{points}.}

\item{formula}{has the form \code{z~x+y}, where \code{z} determines the
bubble sizes and \code{x} and \code{y} determine bubble locations.}

\item{data}{where formula terms are stored, e.g. data frame or list.}

\item{subset}{a logical vector specifying which data to plot.}

\item{na.action}{how \code{NA} values are handled.}
}
\value{
\code{NULL}, but a bubble plot is drawn.
}
\description{
Draw bubble plot, a scatterplot with varying symbol sizes and colors.
}
\note{
Negative \code{z} values are drawn in \code{col[2]} and \code{bg[2]}.
}
