% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Calendar}
\usage{
calendar(filename = NULL, year = 2009, lwd = 0.1, radius = c(1.03,
  0.95, 0.8, 0.65), cex = c(0.3, 0.8, 1.6), col = "white", ...)
}
\arguments{
\item{filename}{EPS output file, or \code{NULL} to draw on screen.}

\item{year}{calendar year.}

\item{lwd}{line width.}

\item{radius}{vector of 4 numbers, setting distance from center to: (1) day
labels, (2) notes boundary, (3) month boundary, (4) year boundary.}

\item{cex}{vector of 3 numbers, setting font size of: (1) day labels, (2)
month labels, (3) year label.}

\item{col}{vector of 4 colors (recycled), setting bg color of: (1) day areas,
(2) notes areas, (3) month areas, (4) year area.}

\item{\dots}{passed to \code{text}.}
}
\value{
Vector of radians (\code{cos} and \code{sin} give day coordinates).
}
\description{
Draw circular calendar with days and months.
}
