% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Uppercase Words}
\usage{
capitalize(x, strict = TRUE)
}
\arguments{
\item{x}{a string (one or more words) to be formatted.}

\item{strict}{whether existing uppercase letters should be lowercased.}
}
\value{
String with every word starting in uppercase.
}
\description{
Make every word in string start in uppercase.
}
\note{
Based on example in \code{help(chartr)}. Better than
  \code{Hmisc::capitalize()}.
}
