% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Change Directory}
\usage{
cd(dir)
}
\arguments{
\item{dir}{an existing directory.}
}
\value{
Working directory before the change.
}
\description{
Set the working directory.
}
\note{
Shorthand for \code{setwd()} for quick navigation in the console.
}
