% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Relative Change}
\usage{
change(from, to = NULL)
}
\arguments{
\item{from}{the original value, or a vector of length 2.}

\item{to}{the new value, or \code{NULL} when \code{from} is a vector of
length 2.}
}
\value{
Relative change between \code{-1} and \code{Inf}, where \code{0} means no
change.
}
\description{
Calculate relative change as a fraction of original value.
}
