% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohorts}
\alias{cohorts}
\title{Transform Table to Cohort Format}
\usage{
cohorts(x, xtab = TRUE, plot = TRUE, log = "", base = exp(1),
  type = "l", lty = 1, xlab = "Age", ylab = "Value", ...)
}
\arguments{
\item{x}{a \code{Year | Age | Value} table (three column data frame or
crosstabbed data frame, matrix, or table).}

\item{xtab}{whether the output should be returned as a wide crosstab table
(otherwise long data frame).}

\item{plot}{whether a cohort plot should be created.}

\item{log}{a string specifying which axes should be log-transformed:
\code{""}, \code{"x"}, \code{"y"}, \code{"xy"}, or \code{TRUE -> "y"}.}

\item{base}{the log base used if \code{log} is \code{TRUE}.}

\item{type}{passed to \code{matplot}.}

\item{lty}{passed to \code{matplot}.}

\item{xlab}{passed to \code{matplot}.}

\item{ylab}{passed to \code{matplot}.}

\item{\dots}{passed to \code{matplot}.}
}
\value{
Crosstab table in cohort format, of class \code{c("xtabs","table")}.
}
\description{
Transform \code{Year | Age | Value} table into cohort format, and possibly
create a plot.
}
\note{
If \code{x} is a three column data frame, column 1 must contain the year,
column 2 the age, and column 3 the value. If \code{x} is crosstabbed, rows
must be years and columns must be age.
}
\section{Warning}{

Zero values are treated as \code{NA}.
}

