% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare.data.frame}
\alias{compare.default}
\alias{compare.function}
\title{Compare Objects}
\usage{
compare(A, B, ...)

\method{compare}{data.frame}(A, B, zoom = TRUE, ...)

\method{compare}{default}(A, B, ...)

\method{compare}{`function`}(A, B, useSource = TRUE, external = FALSE,
  viewer = if (.Platform$OS.type == "windows") "winmerge" else "kompare",
  ...)
}
\arguments{
\item{A}{object to compare.}

\item{B}{object to compare.}

\item{\dots}{passed to \code{compare.*} methods.}

\item{zoom}{whether to show only data frame columns whose values differ.}

\item{useSource}{whether source code format and comments matter, otherwise R
standard format is used.}

\item{external}{whether to compare using an external viewer.}

\item{viewer}{an external program to use if \code{external=TRUE}.}
}
\value{
List with 4 elements:
\item{both}{items found in both objects}
\item{A}{items found only in A}
\item{B}{items found only in B}
\item{same.data}{whether objects contain the same set of items}
}
\description{
Compare two objects as sets.
}
\note{
Set comparison ignores duplicates and order: \code{c(1,1,2,3)} and
\code{c(3,2,1)} are similar. Similarly, when comparing two functions, the
following are the same set:

\preformatted{
f <- function(x)    g <- function(x)
\{                   \{
  y <- 1              y <- 3
  y <- 1              y <- 2
  y <- 2              y <- 1
  y <- 3            \}
\}}

Data frames are compared line by line, and must have the same dimensions.
}
