% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cquantile.R
\name{cquantile}
\alias{cquantile}
\title{Cumulative Quantiles}
\usage{
cquantile(z, probs = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{z}{a numeric vector.}

\item{probs}{a vector of desired quantiles.}
}
\value{
Data frame containing cumulative quantiles, with samples in rows and
probabilities in columns.
}
\description{
Calculate cumulative quantiles.
}
\note{
Implemented and embedded in \code{cumuplot} in the \pkg{coda} package.
}
