% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{whether \code{NA} values should be stripped.}
}
\value{
Coefficient of variation as a single number.
}
\description{
Calculate coefficient of variation.
}
