% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.character.R
\name{diff.character}
\alias{diff.character}
\title{Compare Files}
\usage{
\method{diff}{character}(x, y, ...)
}
\arguments{
\item{x}{a file or folder name.}

\item{y}{another file or folder name.}

\item{\dots}{ignored.}
}
\value{
List containing lines that are different between the files.
}
\description{
Show differences between files or folders.
}
\examples{
x <- system.file("DESCRIPTION", package="base")
y <- system.file("DESCRIPTION", package="datasets")
diff(x, y)

}
