% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot.R
\name{dplot}
\alias{dplot}
\title{Density Plot}
\usage{
dplot(
  ...,
  name = TRUE,
  quantiles = 0.5,
  plot.mean = FALSE,
  main = "",
  xlab = "",
  ylab = "",
  ylim,
  las = 1
)
}
\arguments{
\item{\dots}{a list of values, typically from split(x,factor).}

\item{name}{whether factor names should be displayed on x axis.}

\item{quantiles}{a vector of quantiles where lines should be drawn.}

\item{plot.mean}{whether a thick line should be drawn at each mean.}

\item{main}{passed to \code{plot}.}

\item{xlab}{passed to \code{plot}.}

\item{ylab}{passed to \code{plot}.}

\item{ylim}{passed to \code{plot}.}

\item{las}{passed to \code{plot} and \code{axis}.}
}
\value{
Vector containing subplot baselines on the x axis.
}
\description{
Draw vertical density plots by factor.
}
\author{
Alistair Dunn, with minor changes by Arni Magnusson.
}
