% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{eip}
\alias{eip}
\title{Edit in Place}
\usage{
eip(name, ...)
}
\arguments{
\item{name}{a function, quoted or unquoted.}

\item{\dots}{passed to \code{edit}.}
}
\value{
A copy of the updated function.
}
\description{
Edit function in place.
}
\note{
Based on \code{eip} in package \pkg{Hmisc}.
}
