% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.R
\name{eps}
\alias{eps}
\title{EPS Device}
\usage{
eps(file = "plot.eps", width = 6, height = 4, pointsize = 10,
  horizontal = FALSE, onefile = FALSE, paper = "special", ...)
}
\arguments{
\item{file}{output filename.}

\item{width}{device width in inches.}

\item{height}{device height in inches.}

\item{pointsize}{text size in pts.}

\item{horizontal}{whether to use landscape layout.}

\item{onefile}{whether to produce page output, otherwise EPS output.}

\item{paper}{the kind of paper to print on.}

\item{\dots}{passed to \code{postscript}.}
}
\value{
\code{NULL}, but a graphics device is opened.
}
\description{
Open postscript device for EPS output.
}
\note{
EPS options \code{horizontal}, \code{onefile}, and \code{paper} are explained
on the \code{postscript} help page.
}
