% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps2png.R
\name{eps2png}
\alias{eps2png}
\title{Convert EPS to PNG.}
\usage{
eps2png(epsfile, dpi = 300, gray = FALSE, ag = 4, at = 4)
}
\arguments{
\item{epsfile}{EPS file to convert.}

\item{dpi}{dots per inch.}

\item{gray}{whether image should be grayscale.}

\item{ag}{antialias value for graphics, from 0 (sharp) and 4 (smooth).}

\item{at}{antialias value for text, from 0 (sharp) and 4 (smooth).}
}
\value{
\code{NULL}, but alters EPS file and creates PDF file.
}
\description{
Convert EPS file to PNG format.
}
\note{
Requires shell scripts \code{2png} and \code{optipng}.
}
