% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export Data}
\usage{
export(x)
}
\arguments{
\item{x}{a vector, matrix, table, or data frame.}
}
\value{
Data frame representation of \code{x}.
}
\description{
Print object without quotes or row names.
}
\note{
Useful for copying data from the console to a text editor or spreadsheet
without \verb{[1,]} indicators.
}
