% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdim.R
\name{fdim}
\alias{fdim}
\title{FLR Dimensions}
\usage{
fdim(x)
}
\arguments{
\item{x}{FLR object.}
}
\value{
Named vector showing the dimensions of an FLR object.
}
\description{
Show dimensions of an FLR object.
}
