% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizz.R
\name{fizz1}
\alias{fizz1}
\alias{fizz2}
\title{FizzBuzz}
\usage{
fizz1()

fizz2()
}
\value{
String vector containing numbers, \code{Fizz}, \code{Buzz}, and
\code{FizzBuzz}.
}
\description{
Solutions to FizzBuzz test. \code{fizz1} assigns values to vector subsets,
while \code{fizz2} assigns values to vector elements one by one.
}
\note{
The task is to print the numbers from 1 to 100, with the following
exceptions:
\itemize{
\item for multiples of three print \samp{Fizz} instead of the number
\item for multiples of five print \samp{Buzz}
\item for multiples of both three and five print \samp{FizzBuzz}
}
}
