% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fournier.R
\name{fournier}
\alias{fournier}
\title{Fournier Likelihood}
\usage{
fournier(P, Phat, n)
}
\arguments{
\item{P}{observed catch at age.}

\item{Phat}{predicted catch at age.}

\item{n}{sample size, one number or vector.}
}
\value{
Negative log likelihood.
}
\description{
Evaluate Fournier likelihood, Coleraine-style.
}
