% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\name{ga}
\alias{ga}
\title{Get Anywhere}
\usage{
ga(...)
}
\arguments{
\item{\dots}{a function or function name.}
}
\value{
Shows function.
}
\description{
Show non-visible function.
}
\note{
Shorthand notation for a common task.
}
