% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{whether \code{NA} values should be stripped before the
computation proceeds.}
}
\value{
Geometric mean as a vector.
}
\description{
Calculate geometric mean.
}
