% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github}
\alias{github}
\title{GitHub User}
\usage{
github(user)
}
\arguments{
\item{user}{GitHub username.}
}
\value{
Vector of strings containing \code{@} sign.
}
\description{
Look up email address of GitHub user.
}
