% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{History}
\usage{
h(max.show = Inf, reverse = TRUE, ...)
}
\arguments{
\item{max.show}{the number of previous commands to show (default is entire
log).}

\item{reverse}{whether to show lines in reverse order (not useful when there
are continuation lines).}

\item{\dots}{passed to \code{history}.}
}
\value{
Shows command history in external editor.
}
\description{
Show full command history.
}
\note{
Shorthand notation for a common task.
}
