% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclDemo.R
\name{hclDemo}
\alias{hclDemo}
\title{HCL Colors}
\usage{
hclDemo(n = 400, col = hcl(360 * seq(0, 1 - 1/n, 1/n), c = 120), ...)
}
\arguments{
\item{n}{the number of hue values.}

\item{col}{an optional set of colors, to show some other palette.}

\item{\dots}{passed to \code{hsvDemo}.}
}
\value{
\code{NULL}, but a plot is drawn.
}
\description{
Plot HCL colors as a wheel with \code{c=120} chroma and \code{l=85}
luminance.
}
\seealso{
\code{\link{hsvDemo}}, \code{\link{rgbDemo}}.
}
