% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{heart}
\alias{heart}
\title{Heart}
\usage{
heart(x)
}
\arguments{
\item{x}{value(s) between -2.718 and 2.718.}
}
\value{
Value(s) corresponding to f(x).
}
\description{
Oscillating function.
}
\note{
\deqn{f(x)=\sin(\pi^3x)\times\frac{\sqrt{e^2-x^2}}{2}+\sqrt{|x|}}{
      f(x) = sin(pi^3 * x) * sqrt(exp(2)-x^2)/2 + sqrt(abs(x))}
}
\examples{
x <- seq(-exp(1), exp(1), length=1000)
plot(x, heart(x), type="l", col=2, lwd=2)

}
