% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hess.R
\name{hess}
\alias{hess}
\title{SE from Hessian}
\usage{
hess(H, cor = FALSE)
}
\arguments{
\item{H}{a Hessian matrix.}

\item{cor}{whether to return correlation matrix instead of SE.}
}
\value{
Vector of standard errors, or if \code{cor = TRUE} a correlation matrix.
}
\description{
Evaluate SE or correlation matrix from Hessian matrix.
}
\note{
Covariance matrix can be useful for applying the delta method.
}
