% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2num.R
\name{hex2num}
\alias{hex2num}
\title{Hexadecimal to Number}
\usage{
hex2num(x)
}
\arguments{
\item{x}{a hexadecimal string, like \code{"FF"}.}
}
\value{
Decimal number.
}
\description{
Convert hexadecimal string to number.
}
