% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.R
\name{hockey}
\alias{hockey}
\title{Hockey-Stick Recruitment}
\usage{
hockey(S, Smax, Rmax, eta = 0)
}
\arguments{
\item{S}{spawning biomass.}

\item{Smax}{lowest \code{S} that produces \code{Rmax}.}

\item{Rmax}{maximum recruitment.}

\item{eta}{an optional smoothness parameter, e.g. \code{0.001}.}
}
\value{
Recruitment as a number or vector.
}
\description{
Calculate hockey-stick recruitment.
}
\note{
Smoothness parameter is based on Mesnil and Rochet (2010), but replacing
\verb{gamma^2/4} with \verb{eta*Smax^2}.
}
