% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvDemo.R
\name{hsvDemo}
\alias{hsvDemo}
\title{HSV Colors}
\usage{
hsvDemo(n = 400, col = hsv(seq(0, 1 - 1/n, 1/n)), labels = "n",
  bg = "black")
}
\arguments{
\item{n}{the number of hue values.}

\item{col}{an optional set of colors, to show some other palette.}

\item{labels}{a scale to use as labels: \code{"n"}:none, \code{1}:0-1,
\code{255}:0-255, \code{360}:0-360, \code{"f"}:0-f.}

\item{bg}{the background color.}
}
\value{
\code{NULL}, but a plot is drawn.
}
\description{
Plot HSV colors as a wheel with full saturation and value.
}
\seealso{
\code{\link{hclDemo}}, \code{\link{rgbDemo}}.
}
