% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Data}
\usage{
import(header = FALSE, ...)
}
\arguments{
\item{header}{passed to \code{read.table}.}

\item{\dots}{passed to \code{read.table}.}
}
\value{
Vector or data frame from clipboard.
}
\description{
Read data from clipboard.
}
\note{
Data are read from the clipboard using \code{read.table}, and then converted
to a vector if the data are in a single row or single column.
}
