% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.prime.R
\name{is.prime}
\alias{is.prime}
\title{Prime Numbers}
\usage{
is.prime(n, verbose = FALSE)
}
\arguments{
\item{n}{the number to check.}

\item{verbose}{whether to report the smallest divisor.}
}
\value{
Logical specifying whether number is a prime number.
}
\description{
Test whether number is a prime number.
}
\seealso{
\code{primes} in the \pkg{sfsmisc} package.
}
