% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{l}
\alias{l}
\title{List Objects}
\usage{
l(...)
}
\arguments{
\item{\dots}{passed to \code{ls}.}
}
\value{
Names of objects or elements, as \code{noquote} vector.
}
\description{
Show names of objects or elements.
}
