% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libStatus.R
\name{libStatus}
\alias{libStatus}
\title{Library Status}
\usage{
libStatus(lib.loc = .libPaths(),
  repositories = contrib.url(getOption("repos")), full = FALSE)
}
\arguments{
\item{lib.loc}{a string vector indicating which libraries to analyze.}

\item{repositories}{a string vector indicating which repositories to analyze.}

\item{full}{whether to return a data frame with complete details about all
installed packages.}
}
\value{
List of two data frames,
\item{count}{package count}
\item{details}{packages that are not "ok"}
and a third data frame
\item{full}{all installed packages}
if \code{full = TRUE}.
}
\description{
Summarize status of libraries.
}
