% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.loess.R
\name{lines.loess}
\alias{lines.loess}
\title{Loess Line}
\usage{
\method{lines}{loess}(x, ...)
}
\arguments{
\item{x}{a fitted loess model.}

\item{\dots}{passed to \code{lines}.}
}
\value{
\code{NULL}, but line is added to plot.
}
\description{
Add fitted loess line to an existing plot.
}
