% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lla.R
\name{lla}
\alias{lla}
\title{List All Objects}
\usage{
lla(...)
}
\arguments{
\item{\dots}{passed to \code{ls}.}
}
\value{
Data frame with named rows and the following columns:
\preformatted{
  Class  object class
  KB     object size (name of this column is same as unit used)
  Dim    object dimensions
}
}
\description{
Show information about objects or elements, including hidden ones.
}
\note{
Shorthand for \code{ll(all=TRUE)}.
}
