% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmxy1.R
\name{lmxy1}
\alias{lmxy1}
\title{Simple LM}
\usage{
lmxy1(x, y)
}
\arguments{
\item{x}{a vector containing the predictor values.}

\item{y}{a vector containing the response values.}
}
\value{
Vector containing the coefficients.
}
\description{
Fit a simple LM fast, only returning the coefficients.
}
\seealso{
\code{\link{lmxy2}}.
}
