% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem.R
\name{lorem}
\alias{lorem}
\title{Dummy Text}
\usage{
lorem(n = 1, par = FALSE)
}
\arguments{
\item{n}{the number of paragraphs.}

\item{par}{whether to insert empty string "" between paragraphs.}
}
\value{
Vector of strings.
}
\description{
Create dummy \code{"Lorem ipsum ..."} paragraphs.
}
