% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPass.R
\name{mcmcPass}
\alias{mcmcPass}
\title{MCMC Tests}
\usage{
mcmcPass(x, return.type = "logical")
}
\arguments{
\item{x}{an MCMC trace.}

\item{return.type}{output format.}
}
\value{
Test results in the format specified by \code{return.type}.
}
\description{
Run diagnostic tests on an MCMC trace.
}
\details{
Three \code{return.type} formats are supported:
\tabular{ll}{
  \code{"logical"} \tab a single logical value, indicating whether all tests
                        were passed\cr
  \code{"list"}    \tab a detailed list of all test results\cr
  \code{"string"}  \tab a string indicating which tests failed
}
}
\note{
The test criteria are:
\enumerate{
\item Autocorrelation <= 0.1
\item Geweke Z <= 1.96
\item Heidelberger-Welch p >= 0.05
}

Users can apply other criteria by requesting \code{return.type = "list"}.
}
