% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthsInt.R
\name{monthsInt}
\alias{monthsInt}
\title{Month Number}
\usage{
monthsInt(x)
}
\arguments{
\item{x}{a vector of dates, of class \code{Date} or \code{POSIXt}.}
}
\value{
Vector of months as integers.
}
\description{
Extract month number from a date object.
}
\note{
Use \code{months} to extract month name instead.
}
