% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving.R
\name{moving}
\alias{moving}
\title{Moving Average}
\usage{
moving(x, k = 5, ends = NA)
}
\arguments{
\item{x}{a numeric vector.}

\item{k}{how many values are evaluated in moving average.}

\item{ends}{how the ends should be treated.}
}
\value{
Vector of same length as \code{x}, containing weighted averages.
}
\description{
Calculate moving average.
}
\details{
Alternative \code{end} treatments are:
\tabular{ll}{
  \code{NA}    \tab use \code{NA} values\cr
  \code{FALSE} \tab truncate, so output is shorter than \code{x}\cr
  \code{TRUE}  \tab evaluate moving average on ends, even if the neighbors
                    are all on one side
}
}
\note{
If \code{k} is an even number, a weighted average is used with half weights
on each end.

This function is mainly useful if the ends have flat trends, e.g. a diagonal
line will have bent ends...
}
\seealso{
\code{\link[caTools]{runmean}}.
}
