% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.R
\name{mv}
\alias{mv}
\title{Rename Object}
\usage{
mv(from, to, sure = FALSE, quiet = TRUE)
}
\arguments{
\item{from}{object to rename (quoted or unquoted).}

\item{to}{new object name (quoted or unquoted).}

\item{sure}{whether existing object may be overwritten.}

\item{quiet}{whether report should be suppressed.}
}
\value{
\code{NULL}, but the object is renamed.
}
\description{
Rename object in workspace.
}
