% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.clean.R
\name{na.clean}
\alias{na.clean}
\title{Remove NA Values}
\usage{
na.clean(x)
}
\arguments{
\item{x}{vector or matrix/data frame to check for \code{NA} values.}
}
\value{
Object similar to \code{x}, but possibly shorter (if vector) or fewer rows
(if matrix/data frame).
}
\description{
Remove all elements (if vector) or rows (if matrix/data frame) containing
\code{NA} values.
}
\note{
Similar to \code{na.exclude} and \code{na.omit}, but without special
attributes.
}
