% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano.R
\name{nano}
\alias{nano}
\title{Nano Editor}
\usage{
nano(...)
}
\arguments{
\item{\dots}{is passed to \code{fix}.}
}
\value{
Invisible copy of the updated function.
}
\description{
Edit function with the Nano editor.
}
\note{
To make fast edits without invoking Emacs, when working on a remote Linux
machine.
}
