% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neip.R
\name{neip}
\alias{neip}
\title{Nano Edit in Place}
\usage{
neip(...)
}
\arguments{
\item{\dots}{passed to \code{eip}.}
}
\value{
Invisible copy of the updated function.
}
\description{
Edit function in place, using the Nano editor.
}
\note{
To make fast edits without invoking Emacs, when working on a remote Linux
machine.
}
