% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise}
\alias{noise}
\title{Gaussian Noise}
\usage{
noise(n, mu = 0, sigma = 1, rho = 0)
}
\arguments{
\item{n}{length.}

\item{mu}{mean.}

\item{sigma}{standard deviation (\code{0} to \code{Inf}).}

\item{rho}{autocorrelation (\code{-1} to \code{1}).}
}
\value{
Numeric vector whose mean, standard deviation, and autocorrelation is close
to \code{mu}, \code{sigma}, and \code{rho}.
}
\description{
Generate Gaussian noise with autocorrelation.
}
\note{
Based on ADCAM.
}
\seealso{
\code{\link{rho}}.
}
