% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.like.R
\name{norm.like}
\alias{norm.like}
\title{Normal Likelihood}
\usage{
norm.like(obs, fit)
}
\arguments{
\item{obs}{vector of observed values.}

\item{fit}{vector of fitted values.}
}
\value{
Negative log likelihood.
}
\description{
Calculate the negative log likelihood for a normal distribution.
}
\note{
The negative log likelihood is calculated as:
\deqn{-\log L=0.5n\log(2\pi) + n\log\sigma + \frac{RSS}{2\sigma^2}}{-log
  L = 0.5n log(2 pi) + n log sigma + RSS/(2 sigma^2)}
}
