% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.test.R
\name{norm.test}
\alias{norm.test}
\title{Normal Distribution Test}
\usage{
norm.test(
  x,
  plot = TRUE,
  dots = 50,
  verbose = FALSE,
  pch = 16,
  cex.points = 1,
  ...
)
}
\arguments{
\item{x}{a numeric vector.}

\item{plot}{whether to plot histogram and overlay norm distribution.}

\item{dots}{the number of dots in plotted normal distribution.}

\item{verbose}{whether to return expected curve as a data frame containing
\code{x} and \code{y}.}

\item{pch}{point character.}

\item{cex.points}{point size.}

\item{\dots}{passed to \code{hist}.}
}
\value{
Test result as \code{htest} object.
}
\description{
Test whether sample is normal distributed.
}
\note{
Shapiro-Wilk test is used, as it seems to be preferred over the traditional
G-test and Kolmogorov-Smirnov test.
}
