% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x, method = "uniform")
}
\arguments{
\item{x}{a numeric vector.}

\item{method}{either \code{"uniform"} (range 0-1), \code{"student"} (mean 0,
sd 1), or \code{"relative"} (mean 1).}
}
\value{
Numeric vector.
}
\description{
Normalize vector to a dimensionless scale.
}
\note{
For example, \code{c(10,20,30)} becomes:
\tabular{llll}{
  uniform  \tab 0   \tab 0.5 \tab 1\cr
  student  \tab -1  \tab 0   \tab 1\cr
  relative \tab 0.5 \tab 1   \tab 1.5
}
}
\seealso{
\code{\link{scale}}, equivalent to \code{method = "student"}.
}
