% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.deriv.R
\name{num.deriv}
\alias{num.deriv}
\title{Numerical Derivatives}
\usage{
num.deriv(f, x, delta = 1e-04, align = 0, ...)
}
\arguments{
\item{f}{a function.}

\item{x}{a vector of values to pass to \code{f}.}

\item{delta}{the width of each \code{x} window, in which the slope at
\code{x} will be evaluated.}

\item{align}{alignment of \code{x} window (\code{-1}: left, \code{0}: center,
\code{1}: right).}

\item{\dots}{passed to \code{f}.}
}
\value{
Vector of numerical derivatives.
}
\description{
Calculate numerical derivatives of a univariate function.
}
\note{
Simple finite-difference approximation: \eqn{\Delta f / \Delta x}.
}
