% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2bin.R
\name{num2bin}
\alias{num2bin}
\title{Number to Binary}
\usage{
num2bin(x, width = NULL)
}
\arguments{
\item{x}{is number.}

\item{width}{minimum width for output string.}
}
\value{
Binary string.
}
\description{
Convert number to binary format.
}
