% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2hex.R
\name{num2hex}
\alias{num2hex}
\title{Number to Hexadecimal}
\usage{
num2hex(x, width = NULL, upper = FALSE)
}
\arguments{
\item{x}{a number.}

\item{width}{minimum width for output string.}

\item{upper}{whether to use upper case.}
}
\value{
Hexadecimal string.
}
\description{
Convert number to hexadecimal string.
}
