% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2oct.R
\name{num2oct}
\alias{num2oct}
\title{Number to Octal}
\usage{
num2oct(x, width = NULL)
}
\arguments{
\item{x}{a number.}

\item{width}{minimum width for output string.}
}
\value{
Octal string.
}
\description{
Convert number to octal string.
}
