% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.R
\name{nw}
\alias{nw}
\title{Toggle Emacs}
\usage{
nw()
}
\value{
Resulting \code{getOption("editor")} after adding or removing \verb{-nw}.
}
\description{
Toggle whether Emacs is run with command line option \verb{-nw}.
}
\note{
Useful when working over a slow SSH connection.
}
