% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outGrubbs.R
\name{outGrubbs}
\alias{outGrubbs}
\title{Outlier Test}
\usage{
outGrubbs(x, output = c("stripped", "outliers", "which"), alpha = 0.05,
  log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{output}{the required output, see Value below.}

\item{alpha}{the significance level for a two-way test.}

\item{log}{whether outlier detection should be performed on \code{log(x)}.}
}
\value{
Vector like \code{x} with outliers removed, or if \code{output="outliers"}
the outlier values, or if \code{output="which"} a vector indicating which
elements are outliers.
}
\description{
Detect outliers in normal-distributed sample, using Grubbs test.
}
\note{
Grubbs (1969) cited in Sokal and Rohlf (1995, p. 407), recommending sample
size > 25.
}
