% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outTrim.R
\name{outTrim}
\alias{outTrim}
\title{Trim Outliers}
\usage{
outTrim(x, fraction = 0, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{fraction}{the fraction (0 to 0.5) of observations to be trimmed from
each end of \code{x}.}

\item{na.rm}{whether \code{NA} values should be removed from vector before
trimming.}
}
\value{
Sorted vector, same size or shorter than \code{x}.
}
\description{
Trim extreme values from a vector.
}
