% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.arrows.R
\name{p.arrows}
\alias{p.arrows}
\title{Arrows with Solid Heads}
\usage{
p.arrows(
  x1,
  y1,
  x2,
  y2,
  size = 1,
  width = (sqrt(5) - 1)/4/cin,
  fill = 2,
  border = NULL,
  ...
)
}
\arguments{
\item{x1}{x start coordinates.}

\item{y1}{y start coordinates.}

\item{x2}{x end coordinates.}

\item{y2}{y end coordinates.}

\item{size}{arrowhead size.}

\item{width}{arrowhead width.}

\item{fill}{arrowhead fill color.}

\item{border}{arrowhead border color.}

\item{\dots}{is passed to \code{segments}.}
}
\value{
\code{NULL}, but arrowhead is drawn.
}
\description{
Draw arrows with solid heads.
}
\note{
Borrowed from package \code{sfsmisc}, adding \code{border} argument.
}
