% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.blank.R
\name{pdf.blank}
\alias{pdf.blank}
\title{Blank PDF}
\usage{
pdf.blank(file, width = 8.5, height = 11)
}
\arguments{
\item{file}{the filename to create.}

\item{width}{the page width in inches.}

\item{height}{the page height in inches.}
}
\value{
Invisible string that was written to PDF file.
}
\description{
Create blank PDF page.
}
\note{
Exact size, unlike the \code{pdf} device.
}
