% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loess.R
\name{plot.loess}
\alias{plot.loess}
\title{Plot Loess}
\usage{
\method{plot}{loess}(x, lwd = 2, line.col = "black", xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{a fitted loess model.}

\item{lwd}{the line width.}

\item{line.col}{the line colour.}

\item{xlab}{passed to \code{plot}.}

\item{ylab}{passed to \code{plot}.}

\item{\dots}{passed to \code{plot}.}
}
\value{
\code{NULL}, but a plot is drawn.
}
\description{
Plot loess model fit as a line superposed on scatter.
}
