% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcorr.col.R
\name{plotcorr.col}
\alias{plotcorr.col}
\title{Plot Correlation}
\usage{
plotcorr.col(corr, pow = 4, type = "lower", order.fun = NULL,
  col = "col", ...)
}
\arguments{
\item{corr}{a correlation matrix.}

\item{pow}{a power coefficient to choose colors for weak and strong
correlations.}

\item{type}{which part of the matrix should be plotted: \code{"full"},
\code{"lower"}, or \code{"upper"}.}

\item{order.fun}{a function to reorder variables w.r.t. absolute
correlations: e.g. \code{mean}, \code{median}, or \code{max}.}

\item{col}{either \code{"bw"}, \code{"col"}, or a vector/matrix of colors.}

\item{\dots}{passed to \code{plotcorr}.}
}
\value{
Invisible correlation matrix, reordered if \code{order.fun} was used.
}
\description{
Plot correlation matrix as colored ellipses.
}
\note{
Thin wrapper to pass colors to \code{plotcorr}.
}
