% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{pls}
\alias{pls}
\title{Print Simple List}
\usage{
pls(...)
}
\arguments{
\item{\dots}{passed to \code{print.simple.list}.}
}
\description{
Print simple list.
}
\note{
Shorthand notation for \code{print.simple.list()}.
}
\examples{
pls(options())

}
