% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{psl}
\alias{psl}
\title{Print Simple List}
\usage{
psl(...)
}
\arguments{
\item{\dots}{passed to \code{print.simple.list}.}
}
\description{
Print simple list.
}
\note{
Shorthand notation for \code{print.simple.list()}.
}
\examples{
psl(options())

}
