% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.pdf.stamp.R
\name{remove.pdf.stamp}
\alias{remove.pdf.stamp}
\title{Remove PDF Stamp}
\usage{
remove.pdf.stamp(file)
}
\arguments{
\item{file}{is a PDF document.}
}
\value{
Overwrites original PDF file, after removing stamps.
}
\description{
Replace PDF stamp like \samp{(Downloaded by [Arni Magnusson] ...)} with
spaces.
}
\note{
Similar to Emacs Lisp function \verb{remove-pdf-stamp}.
}
