% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require.R
\name{Require}
\alias{Require}
\title{Require Package}
\usage{
Require(package)
}
\arguments{
\item{package}{package name(s).}
}
\value{
Logical indicating whether the package is ready to be loaded.
}
\description{
Install package if it is not already installed.
}
