% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbDemo.R
\name{rgbDemo}
\alias{rgbDemo}
\title{RGB Colors}
\usage{
rgbDemo(n = 40, cex = 1, bg = "black")
}
\arguments{
\item{n}{the number of levels for each R/G/B component.}

\item{cex}{the size of plot symbols.}

\item{bg}{the background color.}
}
\value{
\code{NULL}, but a plot is drawn.
}
\description{
Plot RGB colors as a triangle.
}
\seealso{
\code{\link{hclDemo}}, \code{\link{hsvDemo}}.
}
