% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Autocorrelation}
\usage{
rho(x)
}
\arguments{
\item{x}{a vector of values.}
}
\value{
Autocorrelation as a number.
}
\description{
Calculate serial autocorrelation with lag 1.
}
\note{
Gives same value as \code{acf(lag=1)}.

In the calculations, the two vectors \code{x[t]} and \code{x[t+1]} are
compared to the same \code{xbar}.

Therefore, the estimated autocorrelation is slightly lower than
\code{cor(x[t], x[t+1])} and also lower than \code{cov(x[t], x[t+1]) /
var(x)}, where \code{x[t]} and \code{x[t+1]} are not compared to a common
mean.

This function serves as a shorthand for \code{acf(x, plot=FALSE,
lag.max=1)$acf[2]} and makes the underlying equation transparent.
}
\seealso{
\code{\link{noise}}.
}
