% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker}
\alias{ricker}
\title{Ricker Recruitment}
\usage{
ricker(S, a, b)
}
\arguments{
\item{S}{spawning biomass.}

\item{a}{initial slope.}

\item{b}{\code{1/Smax}.}
}
\value{
Recruitment as a number or vector.
}
\description{
Calculate Ricker recruitment (Beverton-Holt formulation).
}
\note{
Formulation from Beverton and Holt (1957, Eq. 6.16, p. 56).

Scale the y axis with \code{a}.
}
