% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.Rmax.R
\name{ricker.Rmax}
\alias{ricker.Rmax}
\title{Ricker Recruitment}
\usage{
ricker.Rmax(Rmax, Smax, S)
}
\arguments{
\item{Rmax}{maximum recruitment.}

\item{Smax}{stock size that gives \code{Rmax}.}

\item{S}{spawning biomass.}
}
\value{
Recruitment as a number or vector.
}
\description{
Calculate Ricker recruitment (Bjornsson-Magnusson formulation).
}
\note{
Formulation from Bjornsson and Magnusson (2009, Eq. 2.5, p. 6).

Scale the y axis with \code{Rmax}.
}
