% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{Right Substring}
\usage{
right(string, chars)
}
\arguments{
\item{string}{the whole string.}

\item{chars}{the minimum substring length.}
}
\value{
String containing the right part of \code{string} or the whole string.
}
\description{
Get the right part of a string.
}
