% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundp.R
\name{roundp}
\alias{roundp}
\title{Round Proportions}
\usage{
roundp(x, digits = NULL, seats = NULL, method = "SL", labels = names(x))
}
\arguments{
\item{x}{a vector containing proportions (or votes from an election).}

\item{digits}{a number of decimals to use when rounding proportions, max 6.}

\item{seats}{the number of seats to allocate from election results.}

\item{method}{\code{"DH"}, \code{"MSL"}, or \code{"SL"} (default), indicating
the algorithm to use.}

\item{labels}{an optional vector of labels.}
}
\value{
Numeric vector of same length as \code{x}, with rounded proportions whose sum
is 1 (or integers whose sum is \code{seats}).
}
\description{
Round proportions so they sum to one (or allocate seats from election
results).
}
\note{
This function should be called either with a \code{digits} argument, or with
a \code{seats} argument, not both.

The algorithms were designed to allocate seats from elections:
\tabular{ll}{
  \code{"DH"}  \tab (D'Hondt) favors big parties - used in most of Europe and
                    elsewhere\cr
  \code{"MSL"} \tab (modified Sainte-Lague) favors big parties slightly -
                    used in Norway and Sweden\cr
  \code{"SL"}  \tab (Sainte-Lague) does not favor big or small parties - used
                    in New Zealand, Bosnia, and Latvia
}
}
