% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satcol.R
\name{satcol}
\alias{satcol}
\title{Saturate Colors}
\usage{
satcol(col, s = 1)
}
\arguments{
\item{col}{vector of colors (name, hexadecimal string, or integer).}

\item{s}{saturation level.}
}
\value{
Vector of colors as hexadecimal string.
}
\description{
Adjust color saturation.
}
