% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp.R
\name{sdp}
\alias{sdp}
\title{Standard Deviation in Population}
\usage{
sdp(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or the like.}

\item{na.rm}{whether \code{NA} values should be stripped.}
}
\value{
Standard deviation, as a single number.
}
\description{
Calculate standard deviation based on entire population.
}
\note{
Useful to evaluate MLE of \eqn{\sigma}.
}
