% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.R
\name{sel}
\alias{sel}
\title{Selectivity}
\usage{
sel(Sfull, Sleft, Sright = 15, ages = 1:10)
}
\arguments{
\item{Sfull}{age at full selectivity.}

\item{Sleft}{a shape parameter for the left hand slope.}

\item{Sright}{a shape parameter for the right hand slope.}

\item{ages}{a vector of ages.}
}
\value{
Vector of selectivity values.
}
\description{
Calculate selectivity at age for Coleraine double-normal curve.
}
\note{
The shape parameter for each slope is log variance. Use \code{-15} for
knife-edge and \code{15} for flat selectivity.
}
\references{
Hilborn, R., Maunder, M., Parma, A., Ernst, B., Payne, J., and Starr, P.
(2003).
\emph{Coleraine: A generalized age-structured stock assessment model. User's
manual version 2.0.}
University of Washington Report SAFS-UW-0116.
}
