% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{Shannon Index}
\usage{
shannon(x, bounded = TRUE)
}
\arguments{
\item{x}{a frequency vector, one value per species.}

\item{bounded}{whether to bound the index between 0 and 1.}
}
\value{
Species diversity as a number.
}
\description{
Calculate Shannon index of species diversity.
}
\seealso{
\code{\link{yule}}.
}
