% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showColors.R
\name{showColors}
\alias{showColors}
\title{Show Colors}
\usage{
showColors(cex = 0.6)
}
\arguments{
\item{cex}{point size.}
}
\value{
\code{NULL}, but colors are shown in a plot.
}
\description{
Show all X11 colors.
}
\note{
Based on script presented by Kyle Ambert and Steven Bedrick in BMI 507/607,
Lecture 10 (30 Oct 2009).
}
\seealso{
\code{\link{showPalettes}}.
}
