% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPalettes.R
\name{showPalettes}
\alias{showPalettes}
\title{Show Palettes}
\usage{
showPalettes(palette = NULL, cex = 1)
}
\arguments{
\item{palette}{palette name, or \code{NULL} to show all palettes.}
}
\value{
List of palettes.
}
\description{
Show system palettes.
}
\seealso{
\code{\link{showColors}}.
}
