% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkRownames.R
\name{shrinkRownames}
\alias{shrinkRownames}
\title{Shrink Row Names}
\usage{
shrinkRownames(x, force = FALSE)
}
\arguments{
\item{x}{a data frame.}

\item{force}{whether to overwrite current row names with integers 1, 2, 3,
\dots, n.}
}
\value{
Same data frame, possibly with integer row names.
}
\description{
Convert row names to integers.
}
\note{
The default behavior is to convert row names only if they are strings
representing integers: \code{"1"}, \code{"2"}.
}
