% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify Object}
\usage{
simplify(x)
}
\arguments{
\item{x}{a list, data frame, matrix, or vector.}
}
\value{
Object of same dimensions as \code{x}, using simplest possible storage modes.
}
\description{
Coerce object to the simplest possible storage mode.
}
\note{
\verb{factor -> character -> numeric -> integer}

If user knows that \code{x} is a factor coercible to numeric/integer, then
\code{as.numeric(levels(x))[x]} is faster.

Numbers outside the two billion range are not coerced to integers:
\code{c(-2147483647, 2147483647)}.

Faster than \code{type.convert(as.character(x))}.
}
