% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.data.frame.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort Data Frame}
\usage{
\method{sort}{data.frame}(x, decreasing = FALSE, by = seq_along(x),
  na.last = TRUE, ...)
}
\arguments{
\item{x}{a data frame.}

\item{decreasing}{whether to reverse the sorting order.}

\item{by}{a vector containing column numbers or column names to sort by,
negative means descending.}

\item{na.last}{whether \code{NA} values are treated as being last in the
sorting order.}

\item{\dots}{ignored (S3 template).}
}
\value{
Data frame containing sorted data.
}
\description{
Sort data frame by one or more columns.
}
\note{
Based on an S News email from Xiangyang Liu, 7 Aug 1995.

As pointed out in the \verb{R-FAQ}, \code{x[order(x$a,-x$b),]} can also be
used to sort by multiple columns. That approach, using the minus sign to
reverse the sort order, will not work for string variables unless they are
wrapped inside \code{rank}: \code{x[order(x$a,-rank(x$b)),]}.
}
