% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{splot}
\alias{splot}
\title{Single Quantile Plot}
\usage{
splot(x, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{a vector, matrix, or data frame.}

\item{xlab}{x axis label (only used if \code{x} is a vector).}

\item{ylab}{y axis label (only used if \code{x} is a vector).}

\item{\dots}{passed to plot and pairs.}
}
\value{
Invisible fractions, ordered from 0 to 1.
}
\description{
Single quantile plot, or a scatterplot matrix with quantile plots diagonal
panels.
}
\note{
Somewhat similar to \code{plot.data.frame} in S-Plus.
}
