% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss0.R
\name{ss0}
\alias{ss0}
\title{Sum of Squares}
\usage{
ss0(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{whether \code{NA} values should be stripped.}
}
\value{
Sum of squared values, as a single number.
}
\description{
Calculate sum of squared values, e.g. RSS.
}
\note{
\code{RSS = ss0(resid)}, not \code{ssmean(resid)}.
}
