% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmean.R
\name{ssmean}
\alias{ssmean}
\title{Sum of Squares}
\usage{
ssmean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{whether \code{NA} values should be stripped.}
}
\value{
Sum of squared residuals from the mean, as a single number.
}
\description{
Calculate sum of squared residuals from the mean (not RSS).
}
\note{
\code{RSS = ss0(resid)}, not \code{ssmean(resid)}.

Similar speed in a benchmark with 1e5 calls of length 100:
\enumerate{
\item \code{sumx<-sum(x); output<-sum(x*x)-sumx*sumx/length(x)}
\item \code{output<-.Internal(cov(x,NULL,1,FALSE))*(length(x)-1)}, which is
      the same as \code{var(x)*(length(x)-1)}
\item \code{x_xbar<-x-.Internal(mean(x)); output<-sum(x_xbar*x_xbar)}
}

These are all 3 x faster than \code{output<-sum((x-mean(x))^2)}.

Implementation 1 is used here, since the algorithm can be used in other
languages.
}
