% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmode.R
\name{statmode}
\alias{statmode}
\title{Statistical Mode}
\usage{
statmode(x, all = FALSE, ...)
}
\arguments{
\item{x}{an object, usually a vector or matrix.}

\item{all}{whether to return all modes (default returns only first).}

\item{\dots}{passed to \code{table}.}
}
\value{
The most frequent value in \code{x}, possibly a vector if \code{all = TRUE}.
}
\description{
Get statistical mode, the most frequent value.
}
