% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2df.R
\name{tab2df}
\alias{tab2df}
\title{Table to Data Frame}
\usage{
tab2df(x)
}
\arguments{
\item{x}{a table.}
}
\value{
Data frame with same data as \code{x}.
}
\description{
Convert table (including xtabs) to data frame.
}
