% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2mat.R
\name{tab2mat}
\alias{tab2mat}
\title{Table to Matrix}
\usage{
tab2mat(x)
}
\arguments{
\item{x}{a table.}
}
\value{
Matrix with same data as \code{x}.
}
\description{
Convert table (including xtabs) to matrix.
}
