% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trellis.R
\name{trellis}
\alias{trellis}
\title{Trellis Device}
\usage{
trellis(color = FALSE, ...)
}
\arguments{
\item{color}{whether colors are displayed.}

\item{\dots}{passed to \code{trellis.device}.}
}
\value{
\code{NULL}, but a device is opened.
}
\description{
Open \code{trellis.device} with black and white defaults.
}
\note{
Shorthand for \code{trellis.device(color=FALSE, ...)}.
}
