% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u}
\alias{u}
\title{Update Packages}
\usage{
u(...)
}
\arguments{
\item{\dots}{passed to \code{update.packages}.}
}
\description{
Update packages.
}
\note{
Shorthand notation for \code{update.packages()}.
}
\examples{
\dontrun{
u()
u(libPaths()[1])  # skip core packages
}

}
