% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url2txt.R
\name{url2txt}
\alias{url2txt}
\title{URL to Text}
\usage{
url2txt(x)
}
\arguments{
\item{x}{a percent-encoded URL string, like
\code{"http\%3A\%2F\%2Fwww.hafro.is\%2F\%7Earnima"}.}
}
\value{
Plain text, like \code{"http://www.hafro.is/~arnima"}.
}
\description{
Decode URL to text.
}
\note{
Percent-encoding is defined in RFC 3986
(\url{https://tools.ietf.org/html/rfc3986#section-2.1}).
}
