% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varp.R
\name{varp}
\alias{varp}
\title{Variance in Population}
\usage{
varp(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or the like.}

\item{na.rm}{whether \code{NA} values should be stripped.}
}
\value{
Variance, as a single number.
}
\description{
Calculate variance based on entire population.
}
\note{
Useful to evaluate MLE of \eqn{\sigma^2}.
}
