% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonBert.R
\name{vonBert}
\alias{vonBert}
\title{Von Bertalanffy Growth}
\usage{
vonBert(ages = 1:10, Linf = 100, K = 0.2, t0 = 0)
}
\arguments{
\item{ages}{a vector of ages for which lengths should be predicted.}

\item{Linf}{von Bertalanffy growth parameter.}

\item{K}{von Bertalanffy growth parameter.}

\item{t0}{von Bertalanffy growth parameter.}
}
\value{
Vector of predicted lengths
}
\description{
Evaluate the von Bertalanffy growth function.
}
