% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.test.R
\name{wald.test}
\alias{wald.test}
\title{Wald Test}
\usage{
wald.test(est, se, mu, conf.level = 0.95)
}
\arguments{
\item{est}{a point estimate.}

\item{se}{the standard error of the point estimate.}

\item{mu}{the null hypothesis true mean.}

\item{conf.level}{a confidence level that will be used to evaluate a
confidence interval.}
}
\value{
List containing \code{z}, \code{p}, and \code{confint}.
}
\description{
Perform Wald test.
}
\note{
Related to t-test and z-test, commonly used in nonlinear maximum likelihood
estimation.

Assumes that the estimation error is normal-distributed and the standard
error is known.
}
