% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x0.R
\name{x0}
\alias{x0}
\title{X Intercept}
\usage{
x0(model, se = TRUE)
}
\arguments{
\item{model}{a fitted lm or glm of the form \code{y~x}.}

\item{se}{whether to report the standard error.}
}
\value{
List containing \code{est} (estimated x intercept) and \code{se} (standard
error), or if \code{se = FALSE} a simple number.
}
\description{
Evaluate the x intercept and its SE in a simple LM.
}
\note{
The standard error is evaluated using delta method approximation.

In a logistic glm(family=binomial), the x intercept corresponds to the LD50
  inflection point.
}
\section{Warning}{

In reality, the uncertainty about the x intercept is asymmetric.

The delta method SE approximation is inaccurate when the slope is close to
zero.

Clearly, the x intercept SE is infinity when the slope is not significantly
different from zero.
}

