% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xopt.R
\name{xopt}
\alias{xopt}
\title{Optimal X}
\usage{
xopt(model, se = TRUE)
}
\arguments{
\item{model}{a fitted lm or glm of the form \code{y~x+I(x^2)}.}

\item{se}{whether to report the standard error.}
}
\value{
List containing \code{est} (estimated optimal x) and \code{se} (standard
error), or if \code{se = FALSE} a simple number.
}
\description{
Evaluate the optimal x and its SE in a quadratic LM.
}
\note{
The standard error is evaluated using delta method approximation.

The optimal x is the peak or bottom of the quadratic curve. See p. 54 in
Magnusson (2002).
}
\references{
Magnusson, A. (2002).
\emph{Survival rates of coho (Oncorhynchus kisutch) and chinook salmon (O.
tshawytscha) released from hatcheries on the U.S. and Canadian Pacific coast
1972--1998, with respect to climate and habitat effects.}
M.S. thesis, University of Washington.
}
