% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{years}
\alias{years}
\title{Years}
\usage{
years(x)
}
\arguments{
\item{x}{a vector of dates, of class \code{Date} or \code{POSIXt}.}
}
\value{
Vector of years as integers.
}
\description{
Extract year from Date/POSIXt object.
}
