% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yule.R
\name{yule}
\alias{yule}
\title{Yule's K}
\usage{
yule(x)
}
\arguments{
\item{x}{a vector of labels, e.g. people names or words in a book.}
}
\value{
\eqn{K} as a single value
}
\description{
Calculate Yule's characteristic \eqn{K}.
}
\note{
Large value of \eqn{K} indicates lower diversity.

\eqn{K} is proportional to the probability of any two randomly selected items
to be identical.
}
\references{
Yule, G.U. 1944. \emph{The statistical study of literary vocabulary.}
  Cambridge: Cambridge University Press.

McElduff, F., P. Mateos, A. Wade, and M.C. Borja. 2008. What's in a name? The
  frequency and geographic distributions of UK surnames. \emph{Significance}
  \bold{5}:189-192.
}
\seealso{
\code{\link{shannon}}.
}
