% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z.test}
\alias{z.test}
\title{Z-Test.}
\usage{
z.test(x, mu, sigma, n = length(x), conf.level = 0.95)
}
\arguments{
\item{x}{a vector of data values, or the mean of data values.}

\item{mu}{the null hypothesis true mean.}

\item{sigma}{the standard deviation of data values in the population.}

\item{n}{the number of data values.}

\item{conf.level}{a confidence level that will be used to evaluate a
confidence interval.}
}
\value{
List containing \code{xbar}, \code{z}, \code{p}, and \code{confint}.
}
\description{
Perform z-test
}
\note{
Similar to t-test, but assumes that sigma is known without error.

It is wrong to assume that \code{sigma=sqrt(SS/n-1)} or \code{sqrt(SS/n)},
unless \code{x} is the complete population.

It also makes no sense to perform a z-test if \code{x} is the complete
population, since the true mean is known.

Rarely used in statistical inference, unlike the related \code{t.test} and
\code{wald.test} that are commonly used.
}
