% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{History}
\usage{
h(max.show = Inf, reverse = TRUE, ...)
}
\arguments{
\item{max.show}{the number of previous commands to show (default is entire
log).}

\item{reverse}{whether to show lines in reverse order (not useful when there
are continuation lines).}

\item{\dots}{passed to \code{history} and \code{file.show}.}
}
\description{
Show full command history.
}
\note{
Shorthand notation for \code{history()} with improved support for Emacs ESS.

The base \code{history} function returns an error inside an Emacs session. As
a workaround, this function checks if it is run from inside Emacs and if that
is the case, it calls \code{file.show} to show the \file{.Rhistory} file
(ignoring both \code{max.show} and \code{reverse}).
}
