% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x, method = "uniform")
}
\arguments{
\item{x}{a numeric vector.}

\item{method}{either \code{"uniform"} (range 0-1), \code{"student"} (mean 0,
sd 1), or \code{"relative"} (mean 1, original cv).}
}
\value{
Numeric vector.
}
\description{
Normalize vector to a dimensionless scale.
}
\note{
For example, \code{c(10,20,30)} becomes:
\preformatted{
uniform    0.0  0.5  1.0
student   -1.0  0.0  1.0
relative   0.5  1.0  1.5
}
}
\seealso{
\code{\link{scale}}, equivalent to \code{method = "student"}.
}
