% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPalettes.R
\name{showPalettes}
\alias{showPalettes}
\title{Show Palettes}
\usage{
showPalettes(palette = NULL, cex = 1)
}
\arguments{
\item{palette}{palette name, or \code{NULL} to show all palettes.}
}
\value{
List of palettes.
}
\description{
Show system palettes.
}
\examples{
showPalettes()
showPalettes("Tableau")

f <- function() barplot(rep(1, 9), col=col)
par(mfrow=c(2, 3))
col <- palette.colors(9, "Set 1"); f()
col <- palette.colors(9, "Tableau"); f()
col <- palette.colors(9, "Classic Tableau"); f()
col <- palette.colors(9, "Paired"); f()
col <- palette.colors(9, "Alphabet"); f()
col <- palette.colors(9, "Polychrome"); f()
par(mfrow(c(1, 1))

}
\seealso{
\code{\link{showColors}}.
}
