% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u}
\alias{u}
\title{Update Packages}
\usage{
u(lib.loc = .libPaths()[1], ...)
}
\arguments{
\item{lib.loc}{library to update.}

\item{\dots}{passed to \code{update.packages}.}
}
\value{
\code{NULL}, but packages are updated.
}
\description{
Update packages in user library.
}
\note{
Shorthand for \code{update.packages(.libPaths()[1])}.
}
