% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u}
\alias{u}
\title{Update Packages}
\usage{
u(...)
}
\arguments{
\item{\dots}{passed to \code{update.packages}.}
}
\value{
\code{NULL}, but packages are updated.
}
\description{
Update packages.
}
\note{
Shorthand for \code{update.packages}.
}
